Split synteny code into new module
